requirements.txt
APScheduler==3.1.0
Django==1.9.5
django-cors-headers==1.1.0
djangorestframework==3.3.3
djangorestframework-jwt==1.8.0
funcsigs==1.0.2
futures==3.0.5
MySQL-python==1.2.5
PyJWT==1.4.0
python-dateutil==2.5.3
pytz==2016.4
requests==2.10.0
six==1.10.0
tzlocal==1.2.2
wheel==0.24.0